Added cityName parameter; replaced A/B/C with booleans.
That version of `findCountriesByCity` accepts a `cityName` parameter
and uses it in the `$match`
and `$project` stages of the aggregation pipeline.

I replaced the string literals ("A", "B", "C") with
boolean values (true/false).
You can then post-process the results to filter or alter the list
of cities based on this "cityPriority" field.